requirements-dev.txt
-r requirements.txt
black
flake8
pytest
pytest-cov
pytest-django
tox
